+++ compiler/utils/names.cpp
@@ -60,7 +60,7 @@ void setDefNameProperty(Tree t, Tree id)
 void setDefNameProperty(Tree t, const string& name)
 {
 	//cerr << "setDefNameProperty : " << name << " FOR " << t << "#" << boxpp(t) << endl;
-	int		n = name.size();
+	int		n = (int)name.size();
 	int 	m = (gMaxNameSize>1023) ? 1023 : gMaxNameSize;
 	if (n > m) {
 		// the name is too long we reduce it to 2/3 of maxsize
